Push `ImmutableAttributeFactory` down to `VariantMetadataAdapter`
This commit makes the attribute factory available `VariantMetadataAdapter` and `DefaultComponentMetadataHandler`.
It doesn't add any value by itself, but this change is required by subsequent commits, so that we avoid passing
it to mutable module metadata constructors (which would be extremely cumbersome). Instead this commit chooses
to push this down and the methods which will require the factory will be passed the factory as an argument. This
helps segregating usage, and avoids backing "services" into "data" classes.