Inject experimental features service into maven dependency resolving
This is required for deriving java library variants for libraries that
are published with pom metadata only. This changes the resolution result
for dependency with "runtime" scope. Therefore, the new behavior will
only be active in experimental mode in Gradle 4.x.
The injection is done through the resolution engine in a similar manner
as the attribute factory is injected.

Signed-off-by: Jendrik Johannes <jendrik@gradle.com>