Add `withDescription` to `DependencyResolveDetails`
This commit adds the ability to provide a custom, human readable, reason for the selection of a component.
The reason is not yet available in dependency reports. In fact, if the method is called in a build, dependency
insight is going to fail, because we're not yet able to serialize the reason.