Construct matching table more efficiently
To make attribute matching more efficient, we just assume
that all attributes in the schema will be involved in matching
instead of computing the superset of attributes for all candidates.
This avoids filling a LinkedHashSet on every match and we can instead
reuse the already allocated keyset of the schema.

This requires that all attributes are actually declared in the
schema. Until now this was unfortunately not enforced and the Android
plugin uses some attributes without declaring them. To work around this,
this change adds a wrapper around the attribute factory to catch any
undeclared attributes. This wrapper could later be changed to throw
an error instead.